backend-python-api/requirements.txt
fastapi
uvicorn[standard]
pyodbc
python-dotenv
google-generativeai